backend/requirements.txt
# FastAPI Backend Requirements
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.0

# Vector Database (Qdrant)
qdrant-client>=1.11.0

# OpenAI Agents SDK
openai>=1.66.2
openai-agents>=0.0.3

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# CORS and HTTP
httpx>=0.27.0

# Utilities
python-multipart>=0.0.9

